Agregado archivo de DATOS y otros

+++ superAndes-jdo/src/uniandes/isis2304/superAndes/persistencia/SQLProveedor.java
@@ -48,7 +48,7 @@ public class SQLProveedor {
 	 * @param tipoProveedor - El tipo de proveedor, puede ser 'General' o 'Especifico'.
 	 * @return El número de tuplas insertadas
 	 */
-	public long adicionarProveedor (PersistenceManager pm, int nit, String nombre, int calificacion, String tipoProveedor) 
+	public long adicionarProveedor (PersistenceManager pm, long nit, String nombre, int calificacion, String tipoProveedor) 
 	{
         Query q = pm.newQuery(SQL, "INSERT INTO " + psa.darTablaProveedor() + "(nit, nombre, calificacion, tipoProveedor) values (?, ?, ?, ?)");
         q.setParameters(nit, nombre, calificacion, tipoProveedor);
